fix(代码生成): 高版本MySQL数据库最大长度超出整型范围导致无法保存。

+++ seezoon-generator/src/main/java/com/seezoon/generator/dto/db/DbTableColumn.java
@@ -36,7 +36,7 @@ public class DbTableColumn {
     /**
      * 字符长度
      */
-    private Integer maxlength;
+    private Long maxlength;
     /**
      * 列类型 eg:varchar(64)
      */
